NVIDIA: SAUCE: kernel: arm64: Add "model name" for 64bit tasks
BugLink: https://bugs.launchpad.net/bugs/2075392

Currently we have restriction to display processor model name under
cpuinfo structure for 32 bit tasks only and thus, ubuntu unity center
application or the /proc/cpuinfo does not display processor model name
for 64 bit tasks.

With this we are removing the restriction of displaying model name and
/proc/cpuinfo and ubuntu unity center can display the processor
information to the user for 64 bit tasks.

Signed-off-by: Ninad Malwade <nmalwade@nvidia.com>
Nvidia-Reviewed-on: https://git-master.nvidia.com/r/c/3rdparty/canonical/linux-jammy/+/3185525
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
Tested-by: Kevin Fu <chunhuaif@nvidia.com>
Signed-off-by: Brad Griffis <bgriffis@nvidia.com>
Acked-by: Noah Wager <noah.wager@canonical.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Signed-off-by: Noah Wager <noah.wager@canonical.com>