UBUNTU: SAUCE: iommu: arm-smmu: io-pgtable: Add dma_sync API
BugLink: https://bugs.launchpad.net/bugs/2019240

Add an API to iommu's and pagetables that is similar to the TLB flushing
API, but it instead flushes the memory out to the device.
This is called from the iommu framework because the smmu framework isn't
aware of the whole buffer.
Under the old flushing, if the iommu framework's worst case scenario would
be a buffer mapping to 254 contiguous PTE entries. The old framework would
do 254 dma_sync calls, the new framework syncs all of the contiguous page
table entries at once.
Enable the coherent_walk now that we have the dma_sync API enabled so we no
longer do flushing in the normal io-pgtable map/unmaps.

Signed-off-by: Ashish Mhetre <amhetre@nvidia.com>
Signed-off-by: David Gilhooley <dgilhooley@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-by: Ankur Pawar <ankurp@nvidia.com>
Tested-by: Ankur Pawar <ankurp@nvidia.com>
Signed-off-by: Abhilash G <abhilashg@nvidia.com>
Acked-by: Brad Figg <bfigg@nvidia.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Ian May <ian.may@canonical.com>
Signed-off-by: Brad Figg <bfigg@nvidia.com>