UBUNTU: SAUCE: i2c: Error report injection utility support
BugLink: https://bugs.launchpad.net/bugs/2001557

The HSI error report injection utility is an on-bench debug and test
utility which can be used by customers and SQA to test the end-to-end
error detection and reporting path.

Implement the callback function to integrate with the utility, and allow
injecting I2C HSI related error reports.

Bug 3877863

Signed-off-by: Prasun Kumar <prasunk@nvidia.com>
Reviewed-by: Akhil R <akhilrajeev@nvidia.com>
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>
Signed-off-by: Kartik <kkartik@nvidia.com>
Acked-by: Brad Figg <bfigg@nvidia.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Ian May <ian.may@canonical.com>
Signed-off-by: Brad Figg <bfigg@nvidia.com>