UBUNTU: SAUCE: soc: tegra: Add config variable for platform helper support
BugLink: https://bugs.launchpad.net/bugs/2028400

Platform helper support APIs are available on some of the
core kernel and these APIs are used by out-of-tree modules.
To find out from out-of-tree modules, whether core kernel
support the platform helpers or not, add the CONFIG variables.

The out-of-tree modules will check for this config and then
include/use the APIs provided by the platform helper functions.

Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>
Reviewed-by: Vedashree Vidwans <vvidwans@nvidia.com>
Signed-off-by: Abhilash G <abhilashg@nvidia.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Ian May <ian.may@canonical.com>
Signed-off-by: Jacob Martin <jacob.martin@canonical.com>