NVIDIA: SAUCE: iommu: Don't reserve IOVA when address and size are zero
BugLink: https://bugs.launchpad.net/bugs/2034622

When the bootloader doesn't setup the framebuffers, their address and
size are zero in iommu-addresses property. If we intend to use display
driver in kernel without framebuffer then it's causing display mappings
to fail as IOVA is reserved with size and address as zero.
Fix this by checking for size and address of IOVA and skipping the
reservation if both are 0.

http://nvbugs/4222942

Signed-off-by: Ashish Mhetre <amhetre@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Tested-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Noah Wager <noah.wager@canonical.com>
Signed-off-by: Jacob Martin <jacob.martin@canonical.com>