NVIDIA: SAUCE: tty/serial: serial-tegra: DMA improvements
BugLink: https://bugs.launchpad.net/bugs/2079358

Currently, DMA is started after receiving an RX interrupt. This
can increase the latency while reading the initial data from UART
and can cause RX FIFO overflows. This can be improved by starting
DMA during UART startup. This also eliminates the use of RX interrupt
in DMA mode as the driver only need to use RX Timeout and EORD
interrupt to know if the RX is complete.

Fix this by moving the enabling of the DMA to the UART start-up
instead of after receiving the RX interrupt

Signed-off-by: Kartik <kkartik@nvidia.com>
Nvidia-Reviewed-on: https://git-master.nvidia.com/r/c/3rdparty/canonical/linux-jammy/+/3198199
Tested-by: Kevin Fu <chunhuaif@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Signed-off-by: Brad Griffis <bgriffis@nvidia.com>
Acked-by: Noah Wager <noah.wager@canonical.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Signed-off-by: Noah Wager <noah.wager@canonical.com>